requirements.txt
langchain
ipykernel
langchain_community
pypdf
bs4
arxiv
pymupdf
langchain-text-splitters
faiss-cpu
chromadb
langchain_chroma
python-dotenv
streamlit
langchain_groq
